+++ public/background.js
@@ -1,11 +1,6 @@
-console.log('This is the background page.');
-console.log('Put the background scripts here.');
-
 // eslint-disable-next-line no-undef
-chrome.runtime.onMessage.addListener(async (request, sender, sendResponse) => {
+chrome.runtime.onMessage.addListener(async (request) => {
   if (request.message.action === 'SendBitcoin') {
-    // Create a new tab with options page
-    console.log('SendBitcoin InBackground', request.message.params)
     const params = request.message.params;
     // eslint-disable-next-line no-undef
     chrome.windows.create({
@@ -16,8 +11,6 @@ chrome.runtime.onMessage.addListener(async (request, sender, sendResponse) => {
     });
   }
   if (request.message.action === 'ConnectWallet') {
-    // Create a new tab with options page
-    console.log('ConnectWallet InBackground', request.message.params)
     // eslint-disable-next-line no-undef
     chrome.windows.create({
       type: 'popup',
@@ -27,7 +20,6 @@ chrome.runtime.onMessage.addListener(async (request, sender, sendResponse) => {
     });
   }
   if (request.message.includes('ReturnConnectWalletInfo')) {
-    console.log('ReturnConnectWalletInfo in background: ',request.message)
     // eslint-disable-next-line no-undef
     chrome.tabs.query({ active: true }, function(tabs) {
         // eslint-disable-next-line no-undef
@@ -35,7 +27,6 @@ chrome.runtime.onMessage.addListener(async (request, sender, sendResponse) => {
     });
   }
   if (request.message.includes('ReturnSendBitcoin')) {
-    console.log('ReturnSendBitcoin in background: ',request.message)
     // eslint-disable-next-line no-undef
     chrome.tabs.query({ active: true }, function(tabs) {
         // eslint-disable-next-line no-undef