+++ public/test.js
@@ -1,6 +1,3 @@
-console.log('Content script works!');
-console.log('Must reload extension for modifications to take effect.');
-
 var scriptElement = document.createElement('script');
 
 // eslint-disable-next-line no-undef
@@ -9,35 +6,28 @@ scriptElement.src = chrome.runtime.getURL('injectScript.js');
 document.head.appendChild(scriptElement);
 
 window.addEventListener('SendBitcoin', function (event) {
-  console.log('SendBitcoin InContentScript',event.detail);
   // eslint-disable-next-line no-undef
   chrome.runtime.sendMessage(
     {
       message: {action:'SendBitcoin',params:event.detail},
     },
-    function () {
-      console.log('SendBitcoin sent to background script.');
-    })
+    function () {})
 })
 
 window.addEventListener('ConnectWallet', function (event) {
-  console.log('ConnectWallet  InContentScript');
   // eslint-disable-next-line no-undef
   chrome.runtime.sendMessage(
     {
       message: {action:'ConnectWallet',params:event.detail},
     },
-    function () {
-      console.log('ConnectWallet sent to background script.');
-    })
+    function () {})
 })
 
 // eslint-disable-next-line no-undef
 chrome.runtime.onMessage.addListener(
-  function(request, sender, sendResponse) {
+  function(request) {
     // Handle the message from background.js
     if(request.message.includes("ReturnConnectWalletInfo")){
-      console.log('In content script get ReturnConnectWalletInfo from background',request.message);
       const customEvent = new CustomEvent("ReturnConnectWalletInfo", {
         detail: { message: request.message }
       });
@@ -48,9 +38,8 @@ chrome.runtime.onMessage.addListener(
 
 // eslint-disable-next-line no-undef
 chrome.runtime.onMessage.addListener(
-  function(request, sender, sendResponse) {
+  function(request) {
     if(request.message.includes("ReturnSendBitcoin")){
-      console.log('In content script get ReturnSendBitcoin from background',request.message);
       const customEvent = new CustomEvent("ReturnSendBitcoin", {
         detail: { message: request.message }
       });