+++ src/bitcoin/wallet.ts
@@ -6,9 +6,7 @@ import { ScriptData, Signer, Tap, Tx, ValueData, Word } from '@cmdcode/tapscript
 import { addressToScriptPubKey, bigIntToString, parseStringToBigInt, textToHex, toBytes, toInt26, toXOnly } from './helpers';
 import { bitcoin } from './lib/bitcoin-lib';
 import { Utxo } from '../app/app-context';
-import { hexToBytes } from '../utils/hex-to-bytes';
 import { cleanFloat } from '../utils/clean-float';
-import bs58 from 'bs58';
 
 export function generateWallet(network: any) {
     bip39.setDefaultWordlist('english');
@@ -64,9 +62,7 @@ export function generateNewAddress(rootKey: any, network: any, index: number = 0
         output = payments.output;
     }
 
-    // tentativi codifica sbagliata console.log('PUBLICKEY: 1',account.publicKey,account.publicKey.toString(),' 2 ',account.publicKey.toString("utf8", 0, account.publicKey.length, ' 3 ',new TextDecoder('utf-8').decode(account.publicKey)) )//buffer from buffer or hex to string
-  
-    return { rootKey, account: account ?? rootKey, internalPubkey, address, output, publickey: bs58.encode(account.publicKey)  }
+    return { rootKey, account: account ?? rootKey, internalPubkey, address, output, publickey: Buffer.from(account.publicKey, "hex") }
 }
 
 export const sendTokens = (
@@ -200,11 +196,9 @@ export const sendTokens = (
         vout: vout
     });
 
-    const data = hexToBytes(Tx.encode(txdata).hex);
-    const prefix = 160n;
-    const txsize = prefix + BigInt(Math.floor(data.length / 4));
-
-    const fee = rate * txsize * 2n;
+    const txSizeData = Tx.util.getTxSize(txdata);
+    const vsize = BigInt(Math.floor(txSizeData.vsize * 1.2));
+    const fee = (vsize * rate);
 
     vin = [];
     vout = [];
@@ -398,11 +392,9 @@ export const sendSats = (
         vout: vout
     });
 
-    const data = hexToBytes(Tx.encode(txdata).hex);
-    const prefix = 160n;
-    const txsize = prefix + BigInt(Math.floor(data.length / 4));
-
-    const fee = rate * txsize * 2n;
+    const txSizeData = Tx.util.getTxSize(txdata);
+    const vsize = BigInt(Math.floor(txSizeData.vsize * 1.2));
+    const fee = (vsize * rate);
 
     vin = [];
     vout = [];